Use this code to confirm which of the 3 booleans are failing when running the Sample Auto as it did fail on 2/18/25 at around 8pm.
From looking at the video my guess is that the intake slide is bouncing and therefore is failing the "didReachTarget" boolean, but it could be failing any of those booleans.
(Notice the pattern of pulling the booleans out of the if statement for easier troubleshooting and testing)

If that is confirmed then maybe the tolerance of "1" could be adjusted for the xTolerance, or the Config.turnTolerance may need to be avoided and instead use "reachedTargetXY(double xTolerance, double yTolerance)" to pass in a higher value for y.